requirements.txt
# Appium 2.x compatible requirements
Appium-Python-Client>=3.1.0
pytest>=7.4.3
pytest-html>=4.1.1
selenium>=4.15.2
allure-pytest>=2.13.2
Pillow>=10.1.0
configparser>=6.0.0
